Initial work on dat resource

+++ src/rage/fwTypes.h
@@ -1,9 +1,9 @@
 #pragma once
 #include <cstdint>
 
-typedef uint32_t u64;
+typedef uint64_t u64;
 typedef uint32_t u32;
 typedef uint16_t u16;
-typedef uint16_t u8;
+typedef uint8_t u8;
 
 typedef int64_t int64;
